requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==3.0.2
mysql-connector-python==8.0.31
mysqlclient==2.1.1
psycopg2==2.9.5
python-decouple==3.6
python-dotenv==0.21.0
SQLAlchemy==1.4.42
